+++ fractimation/helpers/fractalAlgorithm.py
